requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.109.1
h11==0.14.0
idna==3.6
inflect==7.0.0
pillow==10.2.0
pydantic==2.6.0
pydantic-core==2.16.1
pypng==0.20220715.0
qrcode==7.4.2
sniffio==1.3.0
starlette==0.35.1
typing-extensions==4.9.0
uvicorn==0.27.0.post1
